+++ lib/nadia.ex
@@ -2027,6 +2027,19 @@ defmodule Nadia do
   end
 
   @doc """
+  Use this method to set a new profile photo for the chat. Returns True on success.
+  """
+  @spec set_chat_photo(integer | binary, binary) :: :ok | {:error, Error.t()}
+  @spec set_chat_photo(Client.t(), integer | binary, binary) :: :ok | {:error, Error.t()}
+  def set_chat_photo(chat_id, photo) do
+    api_request("setChatPhoto", [chat_id: chat_id, photo: photo], :photo)
+  end
+
+  def set_chat_photo(%Client{} = client, chat_id, photo) do
+    api_request(client, "setChatPhoto", [chat_id: chat_id, photo: photo], :photo)
+  end
+
+  @doc """
   Use this method to change the title of a chat. Returns True on success.
   """
   @spec set_chat_title(integer | binary, binary) :: :ok | {:error, Error.t()}