+++ lib/nadia/behaviour.ex
@@ -370,6 +370,8 @@ defmodule Nadia.Behaviour do
               :ok | {:error, Error.t()}
   @callback delete_chat_photo(integer | binary) :: :ok | {:error, Error.t()}
   @callback delete_chat_photo(Client.t(), integer | binary) :: :ok | {:error, Error.t()}
+  @callback set_chat_photo(integer | binary, binary) :: :ok | {:error, Error.t()}
+  @callback set_chat_photo(Client.t(), integer | binary, binary) :: :ok | {:error, Error.t()}
   @callback set_chat_title(integer | binary, binary) :: :ok | {:error, Error.t()}
   @callback set_chat_title(Client.t(), integer | binary, binary) :: :ok | {:error, Error.t()}
   @callback set_chat_description(integer | binary) :: :ok | {:error, Error.t()}
@@ -909,6 +911,8 @@ defmodule Nadia.Behaviour do
                       approve_chat_join_request: 3,
                       decline_chat_join_request: 3,
                       delete_chat_photo: 2,
+                      set_chat_photo: 2,
+                      set_chat_photo: 3,
                       set_chat_title: 3,
                       set_chat_description: 2,
                       set_chat_description: 3,